backend/requirements.txt
fastapi
uvicorn
wikipedia-api
spacy
groq
python-dotenv
pydantic
